chore: convert site from jekyll to readme
I like README resources. I really wanted this resource to be useful both
as a README and a single page.

I removed Jekyyl (which nobody liked) and replaced it with a script that
builds the sit from the README.

This actually improves the content by removing the liquid tags and
making everything normal markdown.